Create README.md
This project contains all the necessary files for running a python project implementing GUI using TKinter module for maintaining a record.  Functions that user can use are :  ADDING RECORD , DELETING RECORD , UPDATING RECORD , SEARCHING FOR A RECORD . There are two more functions ; One is for deleting all the records  and the other is to show all records and letting the user navigate them providing the NAVIGATE keys .  'pythonproject.py' is the file that you need to run on your system .  'phone.bin' is the binary file storing data in binary format .  'finalphone.txt' is the text file storing data in the readable form (that too organized ;in the columnar form )  'deletedentries' is the file  (again binary) which keeps the track of the  records deleted by the user .  THESE FILES ARE USED IN THE MAIN FILE i.e. 'pythonproject.py'